build: fixed docker setup
