+++ .env
@@ -1,6 +1,6 @@
 MONGODB_USER=admin
 MONGODB_PASS=pass
-MONGODB_HOST=localhost
+MONGODB_HOST=mongodb
 MONGODB_PORT=27017
 MONGODB_NAME=artemis-web
